+++ example/lib/player/knight.dart
@@ -8,6 +8,8 @@ import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+enum PlayerAttackType { AttackMelee, AttackRange }
+
 class Knight extends SimplePlayer with Lighting, ObjectCollision {
   double attack = 20;
   double stamina = 100;
@@ -71,11 +73,12 @@ class Knight extends SimplePlayer with Lighting, ObjectCollision {
       actionAttack();
     }
 
-    if (event.id == 0 && event.event == ActionEvent.DOWN) {
+    if (event.id == PlayerAttackType.AttackMelee &&
+        event.event == ActionEvent.DOWN) {
       actionAttack();
     }
 
-    if (event.id == 1) {
+    if (event.id == PlayerAttackType.AttackRange) {
       if (event.event == ActionEvent.MOVE) {
         execAttackRange = true;
         angleRadAttack = event.radAngle;
@@ -233,9 +236,9 @@ class Knight extends SimplePlayer with Lighting, ObjectCollision {
           [
             Say(
               "Look at this! It seems that I'm not alone here ...",
-              Container(
-                width: 50,
-                height: 50,
+              person: Container(
+                width: 100,
+                height: 100,
                 child: FutureBuilder<SpriteAnimation>(
                   future: PlayerSpriteSheet.idleRight,
                   builder: (context, anim) {
@@ -283,5 +286,53 @@ class Knight extends SimplePlayer with Lighting, ObjectCollision {
   Future<void> onLoad() async {
     spriteDirectionAttack = await Sprite.load('direction_attack.png');
     return super.onLoad();
+  }
+
+  Say _getSayExample({PersonSayDirection direction = PersonSayDirection.LEFT}) {
+    return Say(
+      "Look at this! It seems that I'm not alone here ...",
+      personSayDirection: direction,
+      person: Container(
+        width: 100,
+        height: 100,
+        color: Colors.red,
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            Text('Person'),
+            SizedBox(
+              height: 80,
+              width: 80,
+              child: FutureBuilder<SpriteAnimation>(
+                future: PlayerSpriteSheet.idleRight,
+                builder: (context, anim) {
+                  if (!anim.hasData) return SizedBox.shrink();
+                  return SpriteAnimationWidget(
+                    animation: anim.data!,
+                    playing: true,
+                  );
+                },
+              ),
+            ),
+          ],
+        ),
+      ),
+      background: Container(
+        color: Colors.green,
+        width: 150,
+        height: 200,
+        child: Text('Background'),
+      ),
+      header: Container(
+        color: Colors.white,
+        margin: EdgeInsets.only(bottom: 10),
+        child: Text('Header'),
+      ),
+      bottom: Container(
+        color: Colors.white,
+        margin: EdgeInsets.only(top: 10),
+        child: Text('Bottom'),
+      ),
+    );
   }
 }