+++ CHANGELOG.md
@@ -19,7 +19,8 @@
 - [BREAKING CHANGE] change `TextConfig` to `TextPaintConfig`
 - adds method `tilePropertiesBelow()` and `tilePropertiesListBelow()` in GameComponent to access proprieties of the tile set in Tiled.
 - adds method `void onCollision(GameComponent component, bool active)` in `ObjectCollision`. Now you can override this method to listen what Component enter in collision.
-- improvements in `BonfireGame`.
+- improvements in `BonfireGame`
+- improvements in `TalkDialog`.
 
 
 # 1.0.3